- Adds new z-transitioning system - Adds beam-based weapons - Enables background processing of PlaceTemplates proc - Renames spawners.dmi file

+++ code/game/atoms.dm
@@ -141,12 +141,14 @@ Also, the icon used for the beam will have to be vertical and 32x32.
 The math involved assumes that the icon is vertical to begin with so unless you want to adjust the math,
 its easier to just keep the beam vertical.
 */
-/atom/proc/Beam(atom/BeamTarget,icon_state="b_beam",icon='icons/effects/beam.dmi',time=50, maxdistance=10)
+/atom/proc/Beam(atom/BeamTarget,icon_state="b_beam",icon='icons/effects/beam.dmi',time=50, maxdistance=10, var/delete = 1)
 	//BeamTarget represents the target for the beam, basically just means the other end.
 	//Time is the duration to draw the beam
 	//Icon is obviously which icon to use for the beam, default is beam.dmi
 	//Icon_state is what icon state is used. Default is b_beam which is a blue beam.
 	//Maxdistance is the longest range the beam will persist before it gives up.
+	var/list/created_overlays = list()
+
 	var/EndTime=world.time+time
 	while(BeamTarget&&world.time<EndTime&&get_dist(src,BeamTarget)<maxdistance&&z==BeamTarget.z)
 	//If the BeamTarget gets deleted, the time expires, or the BeamTarget gets out
@@ -155,9 +157,10 @@ its easier to just keep the beam vertical.
 
 		dir=get_dir(src,BeamTarget)	//Causes the source of the beam to rotate to continuosly face the BeamTarget.
 
-		for(var/obj/effect/overlay/beam/O in orange(10,src))	//This section erases the previously drawn beam because I found it was easier to
-			if(O.BeamSource==src)				//just draw another instance of the beam instead of trying to manipulate all the
-				qdel(O)							//pieces to a new orientation.
+		if(delete)
+			for(var/obj/effect/overlay/beam/O in orange(10,src))	//This section erases the previously drawn beam because I found it was easier to
+				if(O.BeamSource==src)				//just draw another instance of the beam instead of trying to manipulate all the
+					qdel(O)							//pieces to a new orientation.
 		var/Angle=round(Get_Angle(src,BeamTarget))
 		var/icon/I=new(icon,icon_state)
 		I.Turn(Angle)
@@ -166,7 +169,7 @@ its easier to just keep the beam vertical.
 		var/N=0
 		var/length=round(sqrt((DX)**2+(DY)**2))
 		for(N,N<length,N+=32)
-			var/obj/effect/overlay/beam/X=new(loc)
+			var/obj/effect/overlay/beam/X=new(get_turf(src))
 			X.BeamSource=src
 			if(N+32>length)
 				var/icon/II=new(icon,icon_state)
@@ -196,9 +199,18 @@ its easier to just keep the beam vertical.
 					Pixel_y+=32
 			X.pixel_x=Pixel_x
 			X.pixel_y=Pixel_y
+
+			created_overlays += X
+
 		sleep(3)	//Changing this to a lower value will cause the beam to follow more smoothly with movement, but it will also be more laggy.
 					//I've found that 3 ticks provided a nice balance for my use.
-	for(var/obj/effect/overlay/beam/O in orange(10,src)) if(O.BeamSource==src) qdel(O)
+
+	if(delete)
+		for(var/obj/effect/overlay/beam/O in orange(10,src))
+			if(O.BeamSource==src)
+				qdel(O)
+
+	return created_overlays
 
 
 //All atoms
